+++ src/main/java/io/github/a5h73y/parkour/Parkour.java
@@ -28,6 +28,7 @@ import io.github.a5h73y.parkour.plugin.EconomyApi;
 import io.github.a5h73y.parkour.plugin.PlaceholderApi;
 import io.github.a5h73y.parkour.type.challenge.ChallengeManager;
 import io.github.a5h73y.parkour.type.checkpoint.CheckpointManager;
+import io.github.a5h73y.parkour.type.course.CourseInfo;
 import io.github.a5h73y.parkour.type.course.CourseManager;
 import io.github.a5h73y.parkour.type.kit.ParkourKitManager;
 import io.github.a5h73y.parkour.type.lobby.LobbyManager;
@@ -40,6 +41,7 @@ import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
+import java.util.concurrent.CompletableFuture;
 import java.util.stream.Collectors;
 import org.bstats.bukkit.Metrics;
 import org.bukkit.Bukkit;
@@ -47,7 +49,7 @@ import org.bukkit.plugin.java.JavaPlugin;
 
 public class Parkour extends JavaPlugin {
 
-    public static final String PARKOUR = "parkour";
+    public static final String PLUGIN_NAME = "parkour";
 
     private static final int BSTATS_ID = 1181;
     private static final int SPIGOT_PLUGIN_ID = 23685;
@@ -89,7 +91,7 @@ public class Parkour extends JavaPlugin {
         instance = this;
 
         if (parkourNeedsUpgrading()) {
-            Bukkit.getScheduler().runTaskAsynchronously(this, new ParkourUpgrader(this));
+            upgradeParkour();
             return;
         }
 
@@ -100,7 +102,7 @@ public class Parkour extends JavaPlugin {
         setupPlugins();
 
         getLogger().info("Enabled Parkour v" + getDescription().getVersion());
-        new Metrics(this, BSTATS_ID);
+        submitAnalytics();
         checkForUpdates();
     }
 
@@ -258,24 +260,17 @@ public class Parkour extends JavaPlugin {
     }
 
     private void registerCommands() {
-        getCommand(PARKOUR).setExecutor(new ParkourCommands(this));
+        getCommand(PLUGIN_NAME).setExecutor(new ParkourCommands(this));
         getCommand("paconsole").setExecutor(new ParkourConsoleCommands(this));
 
         if (this.getConfig().getBoolean("Other.UseAutoTabCompletion")) {
-            getCommand(PARKOUR).setTabCompleter(new ParkourAutoTabCompleter(this));
+            getCommand(PLUGIN_NAME).setTabCompleter(new ParkourAutoTabCompleter(this));
         }
 
         String json = new BufferedReader(new InputStreamReader(getResource("parkourCommands.json"), StandardCharsets.UTF_8))
                 .lines().collect(Collectors.joining("\n"));
 
         commandUsages = Arrays.asList(new GsonBuilder().create().fromJson(json, CommandUsage[].class));
-
-//        TODO - Run before release, orders json
-//        commandUsages = commandUsages.stream().sorted(Comparator.comparing(CommandUsage::getCommandGroup)
-//                        .thenComparing(CommandUsage::getCommand))
-//                        .collect(Collectors.toList());
-//        commandUsages.forEach(commandUsage -> System.out.println(commandUsage.getCommandGroup() + commandUsage.getCommand()));
-//        System.out.println(new Gson().toJson(commandUsages));
     }
 
     private void registerEvents() {
@@ -287,12 +282,19 @@ public class Parkour extends JavaPlugin {
         getServer().getPluginManager().registerEvents(new SignListener(this), this);
     }
 
+    /**
+     * Check to see if a newer version exists on Spigot.
+     */
     private void checkForUpdates() {
         if (getConfig().getBoolean("Other.CheckForUpdates")) {
             new ParkourUpdater(this, SPIGOT_PLUGIN_ID).checkForUpdateAsync();
         }
     }
 
+    /**
+     * Check to see if the Parkour needs to upgrade.
+     * @return parkour needs to upgrade
+     */
     private boolean parkourNeedsUpgrading() {
         if (super.getConfig().contains("Version")) {
             double existingVersion = super.getConfig().getDouble("Version");
@@ -300,5 +302,26 @@ public class Parkour extends JavaPlugin {
             return existingVersion < currentVersion;
         }
         return false;
+    }
+
+    private void upgradeParkour() {
+        CompletableFuture.supplyAsync(() -> new ParkourUpgrader(this).get())
+                .thenAccept(success -> {
+                    if (success) {
+                        onEnable();
+                    }
+                });
+    }
+
+    /**
+     * Submit bStats analytics.
+     * Can be disabled through the bStats config.yml.
+     */
+    private void submitAnalytics() {
+        Metrics metrics = new Metrics(this, BSTATS_ID);
+        metrics.addCustomChart(new Metrics.SimplePie("number_of_courses", () ->
+                Integer.toString(CourseInfo.getAllCourseNames().size())));
+        metrics.addCustomChart(new Metrics.SingleLineChart("parkour_players", () ->
+                getPlayerManager().getNumberOfParkourPlayer()));
     }
 }