+++ src/main/java/io/github/a5h73y/parkour/enums/Permission.java
@@ -9,38 +9,40 @@ public enum Permission {
 	PARKOUR_ALL("Parkour", "*"),
 
 	// Parkour Basic
-	BASIC_ALL("Parkour.Basic", "*"),
-	BASIC_CHALLENGE("Parkour.Basic", "Challenge"),
-	BASIC_CREATE("Parkour.Basic", "Create"),
-	BASIC_KIT("Parkour.Basic", "Kit"),
-	BASIC_TELEPORT("Parkour.Basic", "TP"),
-	BASIC_TELEPORT_CHECKPOINT("Parkour.Basic", "TPC"),
-	BASIC_LEADERBOARD("Parkour.Basic", "Leaderboard"),
-	BASIC_SIGNS("Parkour.Basic", "Signs"),
-	BASIC_JOINALL("Parkour.Basic", "JoinAll"),
-	BASIC_COMMANDS("Parkour.Basic", "Commands"),
+	BASIC_ALL(Constants.PARKOUR_BASIC, "*"),
+	BASIC_CHALLENGE(Constants.PARKOUR_BASIC, "Challenge"),
+	BASIC_CREATE(Constants.PARKOUR_BASIC, "Create"),
+	BASIC_KIT(Constants.PARKOUR_BASIC, "Kit"),
+	BASIC_TELEPORT(Constants.PARKOUR_BASIC, "TP"),
+	BASIC_TELEPORT_CHECKPOINT(Constants.PARKOUR_BASIC, "TPC"),
+	BASIC_LEADERBOARD(Constants.PARKOUR_BASIC, "Leaderboard"),
+	BASIC_SIGNS(Constants.PARKOUR_BASIC, "Signs"),
+	BASIC_JOINALL(Constants.PARKOUR_BASIC, "JoinAll"),
+	BASIC_COMMANDS(Constants.PARKOUR_BASIC, "Commands"),
 
 	// Parkour Admins
-	ADMIN_ALL("Parkour.Admin", "*"),
-	ADMIN_READY_BYPASS("Parkour.Admin", "ReadyBypass"),
-	ADMIN_LEVEL_BYPASS("Parkour.Admin", "LevelBypass"),
-	ADMIN_COURSE("Parkour.Admin", "Course"),
-	ADMIN_PRIZE("Parkour.Admin", "Prize"),
-	ADMIN_DELETE("Parkour.Admin", "Delete"),
-	ADMIN_RESET("Parkour.Admin", "Reset"),
-	ADMIN_TESTMODE("Parkour.Admin", "Testmode"),
+	ADMIN_ALL(Constants.PARKOUR_ADMIN, "*"),
+	ADMIN_READY_BYPASS(Constants.PARKOUR_ADMIN, "ReadyBypass"),
+	ADMIN_LEVEL_BYPASS(Constants.PARKOUR_ADMIN, "LevelBypass"),
+	ADMIN_COURSE(Constants.PARKOUR_ADMIN, "Course"),
+	ADMIN_PRIZE(Constants.PARKOUR_ADMIN, "Prize"),
+	ADMIN_DELETE(Constants.PARKOUR_ADMIN, "Delete"),
+	ADMIN_RESET(Constants.PARKOUR_ADMIN, "Reset"),
+	ADMIN_TESTMODE(Constants.PARKOUR_ADMIN, "Testmode"),
 
 	// Create Parkour Signs
-	CREATE_SIGN_ALL("Parkour.CreateSign", "*"),
-	CREATE_SIGN_JOIN("Parkour.CreateSign", "Join"),
-	CREATE_SIGN_FINISH("Parkour.CreateSign", "Finish"),
-	CREATE_SIGN_LEAVE("Parkour.CreateSign", "Leave"),
-	CREATE_SIGN_EFFECT("Parkour.CreateSign", "Effect"),
-	CREATE_SIGN_STATS("Parkour.CreateSign", "Stats"),
-	CREATE_SIGN_LEADERBOARDS("Parkour.CreateSign", "Leaderboards"),
-	CREATE_SIGN_LOBBY("Parkour.CreateSign", "Lobby"),
-	CREATE_SIGN_CHALLENGE("Parkour.CreateSign", "Challenge"),
+	CREATE_SIGN_ALL(Constants.PARKOUR_CREATE_SIGN, "*"),
+	CREATE_SIGN_JOIN(Constants.PARKOUR_CREATE_SIGN, "Join"),
+	CREATE_SIGN_JOINALL(Constants.PARKOUR_CREATE_SIGN, "JoinAll"),
+	CREATE_SIGN_FINISH(Constants.PARKOUR_CREATE_SIGN, "Finish"),
+	CREATE_SIGN_LEAVE(Constants.PARKOUR_CREATE_SIGN, "Leave"),
+	CREATE_SIGN_EFFECT(Constants.PARKOUR_CREATE_SIGN, "Effect"),
+	CREATE_SIGN_STATS(Constants.PARKOUR_CREATE_SIGN, "Stats"),
+	CREATE_SIGN_LEADERBOARDS(Constants.PARKOUR_CREATE_SIGN, "Leaderboards"),
+	CREATE_SIGN_LOBBY(Constants.PARKOUR_CREATE_SIGN, "Lobby"),
+	CREATE_SIGN_CHALLENGE(Constants.PARKOUR_CREATE_SIGN, "Challenge"),
 
+	// Other
 	PARKOUR_LEVEL("Parkour.Level", "99999"),
 	PARKOUR_COURSE("Parkour.Course", "*");
 
@@ -62,6 +64,12 @@ public enum Permission {
 
 	public String getPermission() {
 		return permissionRoot + "." + permissionNode;
+	}
+
+	private static class Constants {
+		private static final String PARKOUR_BASIC = "Parkour.Basic";
+		private static final String PARKOUR_ADMIN = "Parkour.Admin";
+		private static final String PARKOUR_CREATE_SIGN = "Parkour.CreateSign";
 	}
 }