+++ src/main/java/io/github/a5h73y/parkour/upgrade/ParkourUpgrader.java
@@ -5,11 +5,12 @@ import io.github.a5h73y.parkour.other.AbstractPluginReceiver;
 import io.github.a5h73y.parkour.other.Backup;
 import java.io.File;
 import java.io.IOException;
+import java.util.function.Supplier;
 import java.util.logging.Logger;
 import org.bukkit.configuration.file.FileConfiguration;
 import org.bukkit.configuration.file.YamlConfiguration;
 
-public class ParkourUpgrader extends AbstractPluginReceiver implements Runnable {
+public class ParkourUpgrader extends AbstractPluginReceiver implements Supplier<Boolean> {
 
 	private final File defaultFile;
 	private final File playerFile;
@@ -48,7 +49,7 @@ public class ParkourUpgrader extends AbstractPluginReceiver implements Runnable 
 	}
 
 	@Override
-	public void run() {
+	public Boolean get() {
 		parkour.getLogger().info("=== Beginning Parkour Upgrade ===");
 		parkour.getLogger().info(String.format("Upgrading from v%s to v%s",
 				defaultConfig.getString("Version"), parkour.getDescription().getVersion()));
@@ -57,19 +58,19 @@ public class ParkourUpgrader extends AbstractPluginReceiver implements Runnable 
 		Backup.backupNow(true);
 
 		if (!new PlayerInfoUpgradeTask(this).start()) {
-			return;
+			return false;
 		}
 
 		if (!new CourseInfoUpgradeTask(this).start()) {
-			return;
+			return false;
 		}
 
 		if (!new StringsConfigUpgradeTask(this).start()) {
-			return;
+			return false;
 		}
 
 		if (!new DefaultConfigUpgradeTask(this).start()) {
-			return;
+			return false;
 		}
 
 		// the database upgrade has to be done in two steps
@@ -77,18 +78,21 @@ public class ParkourUpgrader extends AbstractPluginReceiver implements Runnable 
 		// initialise the actual databases, then transfer the times back into it
 		DatabaseUpgradeTask databaseUpgrade = new DatabaseUpgradeTask(this);
 		if (!databaseUpgrade.start()) {
-			return;
+			return false;
 		}
 
 		parkour.getLogger().info("Setting up Configs and Database...");
 		parkour.registerEssentialManagers();
 
 		if (!databaseUpgrade.doMoreWork()) {
-			return;
+			return false;
 		}
 
 		parkour.getLogger().info("Parkour successfully upgraded to " + parkour.getDescription().getVersion());
-		parkour.getLogger().info("Please restart the server to start the plugin.");
+		parkour.getLogger().info("The plugin will now start up...");
+		parkour.reloadConfig();
+		parkour.onEnable();
+		return true;
 	}
 
 	public FileConfiguration getDefaultConfig() {